% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psupplementary.R
\name{acinar_cells_go_term_analysis}
\alias{acinar_cells_go_term_analysis}
\title{Clustering of genes, plus GO term analysis of them
(Supp Fig 04, Supp Fig 05)}
\usage{
acinar_cells_go_term_analysis(output_dir = ".", ext = "png")
}
\description{
Clustering of genes, plus GO term analysis of them
(Supp Fig 04, Supp Fig 05)
}
