% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acinar_sce}
\alias{acinar_sce}
\title{Aging acinar cells}
\format{A SingleCellExperiment object with [genes] by [cells].}
\source{
\url{https://www.sciencedirect.com/science/article/pii/S009286741731053X}
}
\usage{
acinar_sce
}
\description{
The variable 'donor_age' (age of donor in years) was used as sequential labels. Taken from publication 
"Single-Cell Analysis of Human Pancreas Reveals Transcriptional Signatures of Aging and Somatic Mutation Patterns"
GSE81547
}
\keyword{datasets}
