% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beta_sce}
\alias{beta_sce}
\title{Developing mouse beta cells}
\format{A SingleCellExperiment object with 40737 genes by 575 cells.}
\source{
\url{https://www.cell.com/cell-metabolism/fulltext/S1550-4131(17)30208-5}
}
\usage{
beta_sce
}
\description{
The variable 'age' (developmental stage) was used as sequential labels. Taken from publication 
"Deciphering Pancreatic Islet β Cell and α Cell Maturation Pathways and Characteristic Features at the Single-Cell Level"
GSE87375
}
\keyword{datasets}
