% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colon_sce}
\alias{colon_sce}
\title{[human?] colon cells}
\format{A SingleCellExperiment object with [genes] by [cells].}
\source{
\url{https://www.cell.com/cell-systems/fulltext/S2405-4712(17)30449-0}
}
\usage{
colon_sce
}
\description{
The sequential labels were derived from unsupervised clustering, using Seurat. Taken from publication 
"Unsupervised Trajectory Analysis of Single-Cell RNA-Seq and Imaging Data Reveals Alternative Tuft Cell Origins in the Gut"
GSE102698
}
\keyword{datasets}
