% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{germ_sce}
\alias{germ_sce}
\title{Developing human female germline cells}
\format{A SingleCellExperiment object with [genes] by [cells].}
\source{
\url{https://www.cell.com/cell-stem-cell/fulltext/S1934-5909(17)30078-4}
}
\usage{
germ_sce
}
\description{
The variable 'time' (age in weeks) was used as sequential labels. Taken from publication 
"Single-Cell RNA-Seq Analysis Maps Development of Human Germline Cells and Gonadal-Niche Interactions"
GSE86146
}
\keyword{datasets}
