% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hesc_sce}
\alias{hesc_sce}
\title{Human embryonic stem cells}
\format{A SingleCellExperiment object with 26178 genes by 1529 cells.}
\source{
\url{https://www.ebi.ac.uk/arrayexpress/experiments/E-MTAB-3929/}
}
\usage{
hesc_sce
}
\description{
The variable 'esc_day' (embryonic day) was used as sequential labels. Taken from publication 
"Single-cell RNA-seq reveal lineage formation and X-chromosome dosage compensation in human preimplantation embryos"
E-MTAB-3929
}
\keyword{datasets}
