% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psupplementary.R
\name{plot_dim_reduction}
\alias{plot_dim_reduction}
\title{Nice plotting of dimensionality reduction}
\usage{
plot_dim_reduction(y_vector, dim1, dim2, labels)
}
\description{
Nice plotting of dimensionality reduction
}
