% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psupplementary.R
\name{plot_selected_cluster_ordering}
\alias{plot_selected_cluster_ordering}
\title{Plot this clustering nicely}
\usage{
plot_selected_cluster_ordering(layout_dt, this_order)
}
\description{
Plot this clustering nicely
}
