% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psupplementary.R
\name{unsupervised_clustering_of_colon_cells}
\alias{unsupervised_clustering_of_colon_cells}
\title{Clustering and psupertime analysis of colon cells
(Fig 1F, Supp Fig 12, Supp Fig 13, Supp Fig 14, Supp Fig 15)}
\usage{
unsupervised_clustering_of_colon_cells(output_dir = ".", ext = "png")
}
\description{
Clustering and psupertime analysis of colon cells
(Fig 1F, Supp Fig 12, Supp Fig 13, Supp Fig 14, Supp Fig 15)
}
