% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mef_sce}
\alias{mef_sce}
\title{Mouse embryonic fibroblasts reprogrammed to neurons}
\format{A SingleCellExperiment object with 22524 genes by 315 cells.}
\source{
\url{https://www.nature.com/articles/nature18323}
}
\usage{
mef_sce
}
\description{
The variable 'time_point' (days since induction) was used as sequential labels. Taken from publication 
"Dissecting direct reprogramming from fibroblast to neuron using single-cell RNA-seq"
GSE67310
}
\keyword{datasets}
